Flytter NyNotat til notat pakke.
- Fjerner notatPdfGenerator fra journalpostService.
- Revereter internal verdi på Identitetsnummer.
