+++ app/src/main/kotlin/no/nav/k9punsj/notat/NyNotat.kt
@@ -1,0 +1,9 @@
+package no.nav.k9punsj.notat
+
+data class NyNotat(
+    val søkerIdentitetsnummer: String,
+    val søkerNavn: String,
+    val fagsakId: String,
+    val tittel: String,
+    val notat: String
+)
