+++ app/src/main/kotlin/no/nav/k9punsj/rest/web/openapi/JournalpostController.kt
@@ -11,7 +11,6 @@ import io.swagger.v3.oas.annotations.security.SecurityScheme
 import org.springframework.web.bind.annotation.*
 import io.swagger.v3.oas.annotations.tags.Tag
 import no.nav.k9punsj.journalpost.JournalpostRoutes
-import no.nav.k9punsj.journalpost.NyNotat
 import no.nav.k9punsj.rest.web.IdentOgJournalpost
 
 @RestController
