+++ app/src/test/kotlin/no/nav/k9punsj/notat/NotatRoutesTest.kt
@@ -4,7 +4,6 @@ import kotlinx.coroutines.runBlocking
 import no.nav.helse.dusseldorf.testsupport.jws.Azure
 import no.nav.k9punsj.TestSetup
 import no.nav.k9punsj.journalpost.JournalPostResponse
-import no.nav.k9punsj.journalpost.NyNotat
 import no.nav.k9punsj.util.postAndAssertAwaitWithStatusAndBody
 import no.nav.k9punsj.wiremock.saksbehandlerAccessToken
 import org.assertj.core.api.Assertions
